Address PR Review and adding E2E test
