+++ api/pkg/transformer/types/table/table_test.go
@@ -2,9 +2,12 @@ package table
 
 import (
 	"errors"
+	"fmt"
 	"net/url"
+	"reflect"
 	"testing"
 
+	upiv1 "github.com/caraml-dev/universal-prediction-interface/gen/go/grpc/caraml/upi/v1"
 	"github.com/go-gota/gota/dataframe"
 	gotaSeries "github.com/go-gota/gota/series"
 	"github.com/gojek/merlin/pkg/transformer/spec"
@@ -34,6 +37,245 @@ func TestTable_New(t *testing.T) {
 
 	table2 := NewTable(&gotaDataFrame)
 	assert.Equal(t, gotaDataFrame, *table2.DataFrame())
+}
+
+func Test_NewFromUPITable(t *testing.T) {
+
+	tests := []struct {
+		name        string
+		tableInput  *upiv1.Table
+		want        *Table
+		expectedErr error
+	}{
+		{
+			name: "simple table",
+			tableInput: &upiv1.Table{
+				Name: "table a",
+				Columns: []*upiv1.Column{
+					{
+						Name: "col1",
+						Type: upiv1.Type_TYPE_DOUBLE,
+					},
+					{
+						Name: "col2",
+						Type: upiv1.Type_TYPE_INTEGER,
+					},
+				},
+				Rows: []*upiv1.Row{
+					{
+						RowId: "1",
+						Values: []*upiv1.Value{
+							{
+								DoubleValue: 1.2,
+							},
+							{
+								IntegerValue: 2,
+							},
+						},
+					},
+					{
+						RowId: "2",
+						Values: []*upiv1.Value{
+							{
+								DoubleValue: 2.4,
+							},
+							{
+								IntegerValue: 4,
+							},
+						},
+					},
+				},
+			},
+			want: New(
+				series.New([]float64{1.2, 2.4}, series.Float, "col1"),
+				series.New([]int{2, 4}, series.Int, "col2"),
+				series.New([]string{"1", "2"}, series.String, "row_id"),
+			),
+		},
+	}
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			got, err := NewFromUPITable(tt.tableInput)
+			assert.Equal(t, tt.expectedErr, err)
+			if tt.expectedErr != nil {
+				return
+			}
+			if !reflect.DeepEqual(got, tt.want) {
+				t.Errorf("NewFromUPITable() = %v, want %v", got, tt.want)
+			}
+		})
+	}
+}
+
+func Test_ToUPITable(t *testing.T) {
+
+	tests := []struct {
+		name        string
+		table       *Table
+		tableName   string
+		want        *upiv1.Table
+		expectedErr error
+	}{
+		{
+			name: "simple conversion",
+			table: New(
+				series.New([]float64{1.2, 2.4}, series.Float, "col1"),
+				series.New([]int{2, 4}, series.Int, "col2"),
+				series.New([]string{"row1", "row2"}, series.String, "row_id"),
+			),
+			tableName: "new_table",
+			want: &upiv1.Table{
+				Name: "new_table",
+				Columns: []*upiv1.Column{
+					{
+						Name: "col1",
+						Type: upiv1.Type_TYPE_DOUBLE,
+					},
+					{
+						Name: "col2",
+						Type: upiv1.Type_TYPE_INTEGER,
+					},
+				},
+				Rows: []*upiv1.Row{
+					{
+						RowId: "row1",
+						Values: []*upiv1.Value{
+							{
+								DoubleValue: 1.2,
+							},
+							{
+								IntegerValue: 2,
+							},
+						},
+					},
+					{
+						RowId: "row2",
+						Values: []*upiv1.Value{
+							{
+								DoubleValue: 2.4,
+							},
+							{
+								IntegerValue: 4,
+							},
+						},
+					},
+				},
+			},
+		},
+		{
+			name: "conversation but no row_id column",
+			table: New(
+				series.New([]float64{1.2, 2.4}, series.Float, "col1"),
+				series.New([]int{2, 4}, series.Int, "col2"),
+			),
+			tableName: "table1",
+			want: &upiv1.Table{
+				Name: "table1",
+				Columns: []*upiv1.Column{
+					{
+						Name: "col1",
+						Type: upiv1.Type_TYPE_DOUBLE,
+					},
+					{
+						Name: "col2",
+						Type: upiv1.Type_TYPE_INTEGER,
+					},
+				},
+				Rows: []*upiv1.Row{
+					{
+						RowId: "",
+						Values: []*upiv1.Value{
+							{
+								DoubleValue: 1.2,
+							},
+							{
+								IntegerValue: 2,
+							},
+						},
+					},
+					{
+						RowId: "",
+						Values: []*upiv1.Value{
+							{
+								DoubleValue: 2.4,
+							},
+							{
+								IntegerValue: 4,
+							},
+						},
+					},
+				},
+			},
+		},
+		{
+			name: "simple conversion containing nil",
+			table: New(
+				series.New([]any{1.2, nil}, series.Float, "col1"),
+				series.New([]int{2, 4}, series.Int, "col2"),
+				series.New([]string{"row1", "row2"}, series.String, "row_id"),
+			),
+			tableName: "new_table",
+			want: &upiv1.Table{
+				Name: "new_table",
+				Columns: []*upiv1.Column{
+					{
+						Name: "col1",
+						Type: upiv1.Type_TYPE_DOUBLE,
+					},
+					{
+						Name: "col2",
+						Type: upiv1.Type_TYPE_INTEGER,
+					},
+				},
+				Rows: []*upiv1.Row{
+					{
+						RowId: "row1",
+						Values: []*upiv1.Value{
+							{
+								DoubleValue: 1.2,
+							},
+							{
+								IntegerValue: 2,
+							},
+						},
+					},
+					{
+						RowId: "row2",
+						Values: []*upiv1.Value{
+							{
+								IsNull: true,
+							},
+							{
+								IntegerValue: 4,
+							},
+						},
+					},
+				},
+			},
+		},
+		{
+			name: "failed due to unrecognized type",
+			table: New(
+				series.New([]any{1.2, 2.4}, series.Float, "col1"),
+				series.New([]bool{true, false}, series.Bool, "col2"),
+				series.New([]string{"row1", "row2"}, series.String, "row_id"),
+			),
+			tableName:   "new_table",
+			expectedErr: fmt.Errorf("type bool is not supported in UPI"),
+		},
+	}
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			got, err := tt.table.ToUPITable(tt.tableName)
+			assert.Equal(t, tt.expectedErr, err)
+			if tt.expectedErr != nil {
+				return
+			}
+			if !reflect.DeepEqual(got, tt.want) {
+				t.Errorf("ToUPITable() = %v, want %v", got, tt.want)
+			}
+		})
+	}
 }
 
 func TestTable_Col(t *testing.T) {
